Backend/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
docopt==0.6.2
gunicorn==20.0.4
httpie==2.1.0
idna==2.9
pipreqs==0.4.10
Pygments==2.6.1
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
yarg==0.1.9
